% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chinaWeather-package.R
\docType{package}
\name{chinaWeather-package}
\alias{chinaWeather-package}
\alias{chinaWeather}
\title{China Weather package.}
\description{
China Weather package.
}
\details{
a visualized package for china Weather
}
\keyword{package}
