% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{drawBackground}
\alias{drawBackground}
\title{Draw the background.}
\usage{
drawBackground(title, date, lang = "zh")
}
\arguments{
\item{title}{the image's title}

\item{date}{the date}

\item{lang}{the language zh or en}
}
\description{
Draw the background.
}
