% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{drawDescription}
\alias{drawDescription}
\title{Draw the description.}
\usage{
drawDescription(data, temp, lang = "zh")
}
\arguments{
\item{data}{daily data}

\item{temp}{the temperature}

\item{lang}{the language zh or en}
}
\description{
Draw the description.
}
