% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chinaWeather.R
\name{filename}
\alias{filename}
\title{define a filename from current date}
\usage{
filename(date = Sys.time())
}
\arguments{
\item{date}{input a date type}
}
\value{
character a filename
}
\description{
define a filename from current date
}
\examples{
filename()
filename(as.Date("20110701",format="\%Y\%m\%d"))
}
\keyword{filename}
