\name{getCityInfo}
\alias{getCityInfo}
\title{Get all of city info}
\usage{
  getCityInfo(lang = "en")
}
\arguments{
  \item{lang}{input a language}
}
\value{
  data.frame city info
}
\description{
  Get all of city info
}
\examples{
getWOEIDByCity()
getWOEIDByCity(lang="en")
getWOEIDByCity(lang="zh")
}
\keyword{language}

