% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{getColors}
\alias{getColors}
\title{match the color with ADCODE99.}
\usage{
getColors(temp, breaks)
}
\arguments{
\item{temp}{the temperature}

\item{breaks}{cut the numbers}
}
\value{
new color vector
}
\description{
match the color with ADCODE99.
}
\keyword{color}
