% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{getColors2}
\alias{getColors2}
\title{match the color with ADCODE99.}
\usage{
getColors2(map, prov, ctype)
}
\arguments{
\item{map}{the map data}

\item{prov}{the province}

\item{ctype}{the color type}
}
\value{
new color vector
}
\description{
match the color with ADCODE99.
}
\keyword{color}
