\name{getWeather}
\alias{getWeather}
\title{Get all of city weather Data.}
\usage{
  getWeather(lang = "en", src = "yahoo")
}
\arguments{
  \item{lang}{input a language}

  \item{src}{input data source}
}
\value{
  data.frame weather data
}
\description{
  Get all of city weather Data.
}
\examples{
getWeather()
}
\keyword{weather}

