% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chinaWeather-package.R
\docType{data}
\name{weather20170602}
\alias{weather20170602}
\title{Dataset for 20170602.}
\format{A data frame}
\source{
\url{http://weather.yahooapis.com/forecastrss?w=2151330}
}
\usage{
weather20170602
}
\description{
A demo dataset.
}
\keyword{datasets}
