% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chinaWeather.R
\name{getWOEIDByCity}
\alias{getWOEIDByCity}
\title{Get WOEID of Yahoo By City Name}
\usage{
getWOEIDByCity(en = "beijing")
}
\arguments{
\item{en}{input a English city name}
}
\value{
integer WOEID
}
\description{
Get WOEID of Yahoo By City Name
}
\examples{
\dontrun{
 getWOEIDByCity()
 getWOEIDByCity(en="beijing")
}
}
\keyword{WOEID}
